build: suppress deprecation warning in internal code
